requirements.txt
aiogram==3.4.1
SQLAlchemy[asyncio]==2.0.25
aiosqlite==0.19.0
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
python-dotenv==1.0.0
pyrogram==2.0.106
tgcrypto==1.2.5
numpy>=1.26.0
rank-bm25>=0.2.2
